Revert "V3: Made some changes"
This reverts commit aef6e11c3270e4ceff3c75d53862bde922d131ad.

+++ v3/src/main/java/com/api/v3/medical_slot/DeleteAllMedicalSlotsService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.api.v3.exceptions.ObjectIsNullException;
 import com.api.v3.physician.Physician;
 import com.api.v3.physician.RetrievePhysicianService;
 
@@ -26,6 +27,7 @@ public class DeleteAllMedicalSlotsService {
                 && e.getPhysician().equals(physician)    
             )
             .toList();
+        if (medicalSlots == null) throw new ObjectIsNullException();
         repository.deleteAll(medicalSlots);
         return ResponseEntity.noContent().build();
     }
